requirements.txt
numpy==1.21.5
Pillow==8.2.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3

